Begin of a parallel path floating point adder implementation in DSLX.
Implementation of a floating point adder that splits the computation
into two parallel paths for lower latency.
Code based on papers by P.M. Seidel and G. Even:
   * "On the design of fast IEEE floating-point adders" (2001)
   * "Delay-optimized implementation of IEEE floating-point addition" (2004)

Tests are adding ~40M unique pairs of floating point numbers and compare
the results with the host (c++) addition (subnormals flushed).
Testing floating point numbers cover interesting ranges of exponents and
mantissas to be exhausitve but also not require to test 2^64 combinations...

Note, this code is _incomplete_, it does not yet provide the full
implementation of the r_path and also takes some shortcuts along the
way. But already a good starting point to review.

The code also sometimes looks more closer to Verilog which for now
is to simplify differential review, but will also be more 'canonical'
DSLX in the future.

PiperOrigin-RevId: 599249559