Add tool to remove identifiers from an IR file.
It is often useful to use ir_minimizer_main to reduce some production code that displays some issue in order to obtain a reproduction sample. Unfortunately if one does this it's possible they could leak confidential information based on the names given to the IR elements or source files remaining after minimization. This tool allows one to strip all that information out of the IR.

This is also useful in for just examining minimized files since it removes a lot of the unneeded or even misleading names that come from the original version of the code.

Usage: remove_identifiers_main IR_FILE
PiperOrigin-RevId: 599673019