[DSLX:bytecode] Better invariants for invocation data in the bytecode.
Fundamentally this CL is all about splitting caller from callee parametric
environment, storing it once we've computed it so we don't need to recompute it
later, and dealing with the fallout of that decision.

Previously it was quite confusing that there wasn't both caller and callee
parametric environments available in the bytecode -- we had to go and re-derive
things at interpretation time that we were holding in our hand at bytecode
emission time.

(I believe this may coincidentally solve for some issues we were seeing with
parametrics but I have to add regression tests to confirm.)

Some specific aspects of the changes (it looks like a larger change than I
think it really is):

* Adds TypeInfo::GetTypeInfoTreeString() for debugging visualization purposes
  -- it shows the tree of instantiations for parameterized invocations with
  caller/callee environments annotated.
* Consolidates AddInvocationCallBindings() and SetInvocationTypeInfo() into a
  single method -- AddInvocationTypeInfo() -- this makes it more clear how the
  derived type information is being set for any given parametric invocation.
* Some more minor assertions, safer interfaces, and checks on invariants along
  the way, e.g. changing Inovcation and Spawn bytecode data to be classes
  instead of a struct so we could observe where the "caller bindings" were
  being accessed and make sure that was appropriate.

PiperOrigin-RevId: 599949879