Fixes bug when an existing schedule is passed to an FDO schedule pass.
If an existing schdule is provided to the FDO pass it will add constraints on
nodes being in the same cylce they started in. Since FDO's main purpose is
to move nodes between cycles you might imagine how this could lead to the
infeasible error @ericastor observed.

This bug was masked by the
MutualExclusion pass frequently deleting the schedule after it modified the IR.

PiperOrigin-RevId: 599949988