Allow ir_minimizer_main to extract subsets of the IR
This add a new `--can_extract_segments` flag (default to off) which tells the minimizer to try ripping out random segments of logic and checking them in isolation. The minimizer will only do this if the example is already sufficiently minimized to provide a reasonable chance of hitting reproducing logic (max of 256 nodes by default, configurable with --extract_segments_limit). This feature requires that the flags `--can_remove_params`, `--can_remove_sends`, and `--can_remove_receives` all be passed and the flags `--preserve_channels` and `--test_llvm_jit` not be present since it alters the API of the minimized example radically.

Currently the extractor will not attempt to pull out any invoke-style nodes, instead considering the simplification to not be possible.

For the extracted logic all register-reads, proc-state element reads, and receives are turned into function params. Tokens are not considered part of the logic dependencies.

This greatly improves the minimization in some circumstances. For example for the recent reassociation error it is able to successfully bring the test case down to the 10 or so nodes that actually comprise a minimal test case, removing the extra 20-ish nodes left-over from the proc channels.

PiperOrigin-RevId: 600616622