[DSLX:fmt] Fix truncation bug, cast to underlying STRONG_INT type out of sync.
At some point in the distant past I thought uint16_t was going to be wide
enough, but it was changed to uint32_t when I realized that didn't make sense,
however there was a lingering cast that remained with the old underlying type.
This switches it to appropriately use `.value()` so we can't get out of sync.

When the number of documents got over 65536 the truncation would grab a DocRef
from very early in the process (by virtue of the arena and the DFS traversal
that autoformatting performs), which is why it would spit out something like a
single token or some very small leaf-ish sized construct.

Fixes google/xls#1275

PiperOrigin-RevId: 601293955