Remove `Has*WithName()` functions in favor of getters that return `std::optional<>`.
There are a few places in the codebase where we check to see if a function/proc/block with a given name exists and then try to get it. Getting functions/procs/blocks by name requires a linear search, so it's better to call the getter and see if it returned something rather than checking for existence before getting. The existing getters return `absl::StatusOr<>` which ideally only return errors in exceptional circumstances, so this change adds getter variants that perform the same search but return a `std::optional<>` instead.

This change also updates a few usages of the getters.

PiperOrigin-RevId: 601595895