Extend arithmetic simplification to cover all `umod`/`smod` by a literal
For non-powers-of-2, we convert `x % y` into `x - (x/y)*y`, and let our existing strength-reduction logic take over to convert the division into a multiply and shift(s).

This optimizes away most of the complexity of the `adler32` example benchmark, as XLS now sees through the `umod` of the first iteration and transparently removes most of the work.

PiperOrigin-RevId: 601876420