[DSLX:fmt] Opportunistic-postcondition-validation for autofmt behind a flag.
Use that flag in the autofmt build rule created for all the DSLX language tests
and unit tests.

As noted in the code comments, some transforms the auto-formatter does requires
a grammatical awareness, and we want to avoid anything that requires tree-based
diffing to keep it simple.

```
  // * we may introduce/remove commas
  // * we may reflow comments
  // * we may turn struct instance attributes implicit if the names line up
  // * (unhandled, non-regexp) we may dedup unnecessarily doubled parentheses
  // * (unhandled, non-regexp) we may remove unnecessary parens e.g. on
  //   conditionals and match tested exprs
```

Several of these we can get a conservative comparison by deleting some set of
characters, but for others we'd flag correct transformations. As a result, we
add the flag so it can be used a helpful debugging/testing aid while not
turning it on universally.

Some constructs like redundant field specification in instantiations (i.e.
`MyStruct{x: x}`) could be detected with a backtracking regexp but since we use
RE2 (a non-backtacking regexp engine) these are also not in the set we can
opportunistically check for.

Fixes google/xls#1283

PiperOrigin-RevId: 602796841