[DSLX:ast] Make StructMember entity, makes StructDef more uniform vs EnumDef.
This makes StructDefs more uniform vs EnumDefs which will help with the member
formatting strategy being shared in the DSL formatter.

Note that the NameDef node in the struct members that was previously there was
unnecessary, there's no reason to have def/use chains in the local module unit
for struct members and the use chains weren't populated anyway.

PiperOrigin-RevId: 604700449