Add support for `next_value` nodes to codegen
We remove codegen support for `next (...)` lines with values at the same time, in order to keep from adding a large amount of complexity to codegen. To temporarily maintain compatibility with previous IR, we modernize all procs to use explicit `next_value` nodes before scheduling as part of the state legalization pass.

PiperOrigin-RevId: 604837541