[DSLX:cleanup] Make NameDef member on Import be a ref to document non-nullness.
This is one of those cleanup categories that's hard to switch over "all at
once", so taking the scout approach "leave it better than you found it"
mentality can be useful.

This was better than documenting the "must not be null" property that
corresponded to the XLS_CHECK in the implementation while it was topical.

PiperOrigin-RevId: 604838672