Refactor scheduling passes to work on multiple FunctionBases.
Previously, SchedulingUnit was templated to either take a package or a FunctionBase as the element being operated on, and multiple places baked in the assumption that only one FunctionBase would be scheduled at a time.

This change refactors SchedulingUnit to always take a package and to present a consistent API that exposes a list of FunctionBase* elements to schedule. Passes are updated to operate on each schedulable function.

This is work towards #950 to support multi-proc codegen.

PiperOrigin-RevId: 605027978