[DSLX:TS] Error if we try to resolve colon-ref subject that is not a module.
This gives a proper error if in the expression:

	foo::bar::baz

It turns out that `foo` is not a module. Previously we had a `RET_CHECK`
instead of a proper type inference error if it was not the case.

This showed that the helper routines for resolving the nested ColonRef for type
checking should not live in `ast_utils` as those are parse/AST level helpers,
we want to layer on top of those in `deduce_utils` so we have
appropriately-layered access to errors definitions like `TypeInferenceError`.

Fixes google/xls#1279

PiperOrigin-RevId: 606691090