[DSLX:TS] Add invariant checking for invocation data.
The set of keys in a caller ParametricEnv should always be a subset of the keys
that the caller function has as parametric bindings.

For example, for a function:

	fn p<A: u32, B: u32, C: u32>()

An appropriate `ParametricEnv` would have a subset of {A, B, C} as keys, any
other keys would indicate an invalid parametric env for that function.

PiperOrigin-RevId: 607078254