Faster loop unrolling.
Setting a z3_rlimit will cause the solver to give up on earlier
iterations of the loop and more quickly find later, trivially provable
break conditions.

The potential cost is that some loops with complex break conditions
may fail to unroll, or get extra iterations in the IR/RTL.

However, it is considered best practice to include a simple loop condition,
in the form "for(int i=0;i<MAX_ITERS;++i)", and then use dynamic breaks.

PiperOrigin-RevId: 607440210