[DSLX] Some more modern ref style compliance.
No meaningful functionality changes, should be all cosmetics around refs and
some RET_CHECKs to help support the assertion the pointers were non-null before
they had to be turned into refs, and some clang-tidies that shouted along the
way.

Started asking myself "can this thing be null" and pulled on the thread of "let
me document that it can't" and this is where I ended up. :-)

PiperOrigin-RevId: 607560250