Add checks for codegen metadata.
`StreamingIOPipeline` has a lot of metadata on the side with `Node*` pointers that may be invalidated by codegen passes, in particular DCE. Common examples include valid signals that are literal 1s (common for convenience where each stage takes the previous as an input and the first is set to literal 1) or zero-width ports that are removed.

Unfortunately, adding these checks causes many targets to fail. This change also makes some fields `std::optional<Node*>`[^1] instead of `Node*` and will `reset()` or remove pointers where they can be invalidated.

Work towards #1300.

[^1]: Note that we use `std::optional<Node*>` instead of the more common `nullptr` because our style guide requires bare `Node*` pointers in more places than is typical. See our style guide for more discussion.

PiperOrigin-RevId: 608717012