[DSLX:TS] Invariant that invocation is within associated caller function.
With the code appropriately conforming to the invariant checks this squashes a
few outstanding bugs.

This points out the places where parametric expressions had invocations that
had ambiguous caller functions.

Consider a case like:

```
fn p<X: u32, Y: u32 = {std::clog2(X)}>() { ... }
-----------------------^~~~~~~~~~~~^ invocation of interest

fn main() { p<u32:42>() }
```

The `clog2` call conceptually is happening due to the invocation in `main()`,
but `main` is not really the caller of `clog2` in a sense, `p` is.

As a result, when we go to evaluate parametric default-expressions, we should
push a frame onto the type inferencing stack that says we're within the invoked
function (e.g. `p` in the case above).

Fixes google/xls#1197

PiperOrigin-RevId: 608983091