[DSLX:frontend] Create `ContainedWithinFunction` predicate in `ast_utils.h`
This is a more clear / less fragile way to check that invocations are contained
with in a given function vs using lexical positions (as was previously done).

Note that this originally created a circular dependency between type_info and
ast_util, but that was resolved by moving all helpers that require type
information to deduce_utils instead, which creates more appropriate layering
(AST utilities that just need the AST, vs AST-oriented utilities that also need
type information as determined by type inferencing).

The main changes that are not movement in this CL are:

1. one new routine that does the parent link traversal
2. the place where we make a fake invocation needed a special SetParent setter,
   since materializing AST nodes out of the ether is not something we generally
   do

PiperOrigin-RevId: 609175232