ir_minimizer_main should not consider a pass erroring out a fatal error
ir_minimizer_main is sometimes used to detect errors in passes. It doesn't really make sense to consider a pass failing to run on the minimizing ir to be a fatal error since it might be the issue we are trying diagnose. Instead just consider this to be a failed minimization attempt.

PiperOrigin-RevId: 610447731