Fix issue with bitslice_simp creating illegal selects.
If a dynamic bitslice was determined to be eligable for conversion to select (due to having a 'scaled' start operand) and the start operand has too few bits to address all bits of the input the pass would generate a select with unaddressable cases. This is not a valid IR and would cause the verifier to reject it. Change the pass to trim the select values to an appropriate number in these cases.

Fixes: https://github.com/google/xls/issues/1319
PiperOrigin-RevId: 610517058