Ensure register chaining handles loopbacks correctly
We were incorrectly not taking into account that after the stage where a loopback register is written it does not have the same value anymore. This would mean that the chain needs to be split at some point. Unfortunately the actual resulting chains are very dependent on the exact order the chain is constructed.

PiperOrigin-RevId: 612632136