[DSLX:ir_convert] Rename error from ConversionError to "IrConversionError".
This should be a more helpful indicator of where things are going wrong, vs not
indicating what kind of conversion. Software does lots of conversions in lots
of places. :-) This one is about DSLX to XLS IR.

Note that generally users should never see these except as a form of internal
or unimplemented error, they are only in the exceptional cases where e.g. the
frontend can express something we can't yet convert to IR.

PiperOrigin-RevId: 612913122