[XLS:examples] Move adler32 example into subdirectory.
Normally we don't dote very much on directory structure, but for examples where
users will actively peruse we can give the structure a little extra TLC.

Breaks out adler32 as a first example to a subdirectory because it's one of the
ones that has several files associated with it. Single-file/single-target
examples we may choose to keep flat against the examples directory.

PiperOrigin-RevId: 613038133