[DSLX] Add `zip` builtin.
Note: this is a binary operation to avoid unnecessarily flexing the superpowers
of builtins (e.g. ability to be variable arity), but if n-ary zip was useful
enough we could consider extending it, those kinds of superpowers are nominally
what built-ins can be useful for (though they do make userspace functions feel
bad that they can't do such things).
PiperOrigin-RevId: 613341637