Don't create state elements for context input and output of pipelined loops in FSM mode.
- Context in is already calculated from earlier states' elements.
- Latch context out into the loop body's existing state elements for variables changed.

In order for this to work, the pipelined loop must keep its context output
steady when it is not active in order for later states to use it directly.

Its context output must also be available immediately, in the same tick/activation,
when it is active, in order for nested loops to work without any throughput penalty.

Therefore, the context output is driven by the loop body's state elements by default, and driven by the loop body logic when new values are computed.

There is also some minor clean-up and simplification of the loop body logic.

Also:
- Change the format of the unit test trace output to be more readable
- Add more pipelined loop tests
- Add runtime flags for IR trace generation
PiperOrigin-RevId: 613630299