[DSLX:formal] Add prove_quickcheck_main binary.
Gives us an easy facility to try to prove quickcheck properties from the
command line, building on the XLS SMT translation facilities.

Note: right now this doesn't grab out a counterexample, just provides an error
code / does not display proven. That is an important followup, but it was
landable in MVP form this way.

Fixes google/xls#1337

PiperOrigin-RevId: 613752144