Initial register sharing
This adds a new register sharing optimization to codegen. If enabled registers which share a single value in consecutive, mutually exclusive stages will be merged.

This optimization is controlled with the new `--register_merge_strategy={identity,none}` flag.

By default this is enabled since it should have minimal effect on delay and improve area substantially.

PiperOrigin-RevId: 613983771