Share state elements between top proc and pipelined loops in FSM mode.
A single multimap is accumulated of all the next state values with their conditions.
If all the next conditions for a state value are false, then the state element does not change value.

If only one is true, then that next value will be the next value of the state element, as in the case of values from different FSM states, such as different pipelined loop bodies.

If next conditions are true, then the next value with the lowest "priority" integer becomes the next value of the state element; next conditions with the same priority must be mutually exclusive, and an assertion is generated to ensure this. This is for the case of nested loops, in which the outer most scope should produce the next value for a state element. This value will reflect the logic of the more inner loops, since their context outputs immediately (within the same tick/activation) output the new value they computed this iteration, and that will feed into the logic of the outer loop(s) to drive their next value.

PiperOrigin-RevId: 613987176