[DSLX:run_routines] Refactor into subdir.
There are enough things that are specific to running DSLX stuff as command line
library assistance that it makes sense to refactor these "runner entry point
support" routines into a subdir.

Also adds a README to the DSLX dir just for helpful indicators/navigation for
implementors -- there's not a lot of dirs to avoid overly doting on directory
structure, but for the few we have we can say why we made them.

PiperOrigin-RevId: 614098030