[DSLX:LS] Remove range formatting for now.
We have the new auto-formatting facility and we'd want to use that, but also we
need to figure out what AST construct to rewrite and how to snip the
non-formatted portion around the autoformatted section, and we need to decide a
policy on what to do if there's no an exact collision (do you take the
next-largest scope, closest-fit construct, etc).

For now we'll just remove it as the whole-file formatting is what we typically
use, and we can add back the ranged facility as needed/desired in the future.

PiperOrigin-RevId: 615122216