Start removing old log implementation artifacts.
We switched to absl logging, so we don't need our older log implementation
anymore; in fact it is largely abandoned already, just delete the
old code.

This is a first step in a series of refactoring steps: remove targets
not needed anymore directly.

Upcoming: change the users of remaining logging to directly use
absl logging and remove the targets in this directory.
PiperOrigin-RevId: 615851421