Update clang-tidy script to current environment.
With the clang toolchain, the compilation database
now includes -stdlib=libc++ in the commands, which confused
clang-tidy. Edit compilation database to fix that.

Also: Don't include xlscc/synth_only c++ code to be clang-tidy'ed
That code there is more a low-level library outside of the
usual rules.

.clang-tidy config: ignoring of a new diagnostic message
not relevant with bazel (redefining of system macro).

PiperOrigin-RevId: 615884389