FSM state merging for higher throughput (--merge_states). This allows IO ops to go in the same state as a pipelined loop. Only one pipelined loop can go in a state, however.
By default, this will also split FSM states (--split_states_on_channel_ops) to avoid putting operations on the same channel in the same state. This should avoid the need to use the IO adapter, at the expense of some throughput. It also does not check mutual exclusion, so turning it off may produce better throughput without the IO adapter if the ops are mutually exclusive.

- Also a fix to broken logic when sequential pipelined loops share state elements. For now, any state element that drives a context output for a loop cannot be reused.

- Removed support for ASAP scheduling options for now. This is extra complexity that isn't used. Will add them back once the pipelined loop features have settled down.

PiperOrigin-RevId: 616882759