Add support for AbstractNodeEvaluator to track compound data.
Compound data types (tuples and arrays) are tracked through LTTs of their underlying abstract representation. Due to the absurdly expensive nature of a fully general form of ArrayIndex and ArraySlice using only abstract-evaluator bit-operations these operations are not supported. Other values though are fully tracked and their values can be obtained using the GetCompoundValue function.

The booleanifier has been updated to use the new AbstractNodeEvaluator (since it relied on being able to type-pun arguments in a way that is no longer supported with this change).

PiperOrigin-RevId: 620279261