Track ternary values through arrays and tuples
Arrays and tuples were considered black-boxes through which no ternary information could flow. This could lead to some optimization passes missing things due to lacking information. This lets the TernaryQueryEngine track values through these compound data types.

Note this does not update TernaryQueryEngine to give ternary values for nodes with non-bits type. This is mostly due to it not being clear what this would be useful for. It would be relatively simple to add if desired however.

PiperOrigin-RevId: 620376750