Make sure original source files are available to later rules
This is needed for optimal error reporting with remote-execution builds; otherwise, the source files may not be present for later tools, and `loc`s that reference them won't be able to be resolved.

With this fixed, `PrintCaret` should work even in remote-execution contexts, assuming the path used to reference the original source file is consistent. (In most remote-execution environments, this means using a path relative to the current working directory when possible.)

PiperOrigin-RevId: 621201768