[DSLX:TS] Create AstEnv type wrapper.
This helps call out the places where we're using AST nodes instead of strings
as keys in parametric environments -- it's largely just used on the
`TypecheckInvocation` path and I think that's just some special case kludge for
proc functions.

Still investigating the best way to rework for uniformity but this helped to
pin down the places the AST-based constexpr environments were flowing.

PiperOrigin-RevId: 621310567