Make ternary-iterator match IR behavior around zero-bit values
The IR considers all zero-bit literals to have a single value which is equal to zero. The ternary_ops::AllBitsValues however considered a zero-bit ternary to have no values associated with it. This mismatch could cause some issues with analyzing ir files. Update ternary_ops::AllBitsValues to consider a zero-bit ternary to have a single zero-bit value it can take.

PiperOrigin-RevId: 621549756