Make `InstantiationPath` generic and specialize to `Proc` and `Block`.
The types for `Proc` and `Block` paths are specializations of a common templated type. The idea is that paths are so similar they can easily share an implementation, but it is convenient to specialize them for `Proc` and `Block` cases, especially because the instantiation types for `Proc` and `Block` differ and do not benefit much from being unified. Instead, we define `Instantiated()` and `InstantiatedName()` functions that provide a uniform interface.

PiperOrigin-RevId: 622058265