[DSLX:TS] Report two explicit parametric instantiation errors in stable order.
To make sure we do this properly/consistently we plumb down the full set of
ParametricBindings from the struct or callee-function AST node down into the
parametric instantiation process, and we know that the explicit parametrics
fill in a prefix of these parametric bindings. By iterating the stable order
given there, we know we always get a deterministic traversal order, and
conveniently we also now have the parametric bindings available in the
parametric instantiator.

PiperOrigin-RevId: 622312080