If only one bit needs to be OR'ed, only do that instead of OR with wider type.
This is essentially manual application of
https://github.com/google/xls/issues/1217 and can be removed
once that is implemented in the optimizer.

This is equivalent and synthesis-tools will most likely
figure this out themselves. However, it does help XLS, as
it provides more fine-grained data dependencies.
In this case XLS can distribute logic better into pipeline stages and
reduces flops.

PiperOrigin-RevId: 622315054