Add interval_ops::FromTernary to convert from a ternary into intervals
The IntervalSet FromTernary creates is limited to only consider a limited number of not trailing unknown bits to avoid creating interval sets which (while perfectly preserving the value of the ternary) are to large to be useful. By default up to 16 intervals (4 unknown non-trailing bits) are generated.

Rewrite the QueryEngine::GetIntervals function to use this instead of doing it manually.

PiperOrigin-RevId: 623261446