Add truncate/extend support for InlineBitmap.
Since these are used to implement Bits and similar things this functionality is only exposed as creating a new bitmap with the given size (or operating on an RValue).

Update bits_ops::Zero/SignExtend to work using this and add a new Truncate.

Micro-benchmarks show these as being an 75% less time.

PiperOrigin-RevId: 623290302