Move MinimizeIntervals into interval_ops, add tests and rewrite
The original version used indexes into an array to keep track of the original ordering. This rewrites it to just use an intrusive list which is much clearer in this case.

This new implementation is also more easily modified to support other ranking functions for how to choose which elements to merge since we can just change the heap compare function (and potentially re-run heapify).

PiperOrigin-RevId: 623512492