Add a z3 verifier for proc behavior.
Similar to the function equivalence add a scoped equivalent assert for procs.

This works by unrolling the proc and collecting its sent data (and optionally its state elements). The proc is turned into a function that returns these collected outputs.

There are still several restrictions that might be loosened in the future including that the set of state elements not be changed if these are being returned and that the set of sends not be changed.

It currently relies on each channel only being sent on exactly once in an activation (even if some of the sends are predicated).

Bug: https://github.com/google/xls/issues/1375
Bug: https://github.com/google/xls/issues/1374
PiperOrigin-RevId: 623850650