[DSLX:ir_convert] More defense-in-depth vs FunctionBuilder-error based crashes.
Conceptually the type system is supposed to only allow programs through that
will IR convert correctly, but if we make a mistake and pass a program through
that flags issues in IR conversion, it's nice to not crash and instead give an
indication of the kind of error that occurred.

Towards google/xls#1223 -- the IR builder in this issue flags that the value to
be selected has a token inside which is unhandled by one-hot-select, but we
crash because we look at the BValue's type which goes through the nullptr node.

Conceptually having an "eager error" wrapper around the FunctionBuilder would
probably be preferrable for the IR converter to avoid any latent issues with
error laziness leading to bad nullptr handling.

PiperOrigin-RevId: 623888820