Support instantiations in block interpreter.
Now that we have block elaboration and visitors on those elaborations, we implement support for instantiations in the block interpreter. The idea is that there's a top-level elaborated block visitor that iterates through the elaborated nodes, implements inter-block ops, and then delegates the other ops to a per-block IR interpreter.

These changes also extend to the block evaluator, which needs to elaborate in some places, update how registers are initialized, etc.

PiperOrigin-RevId: 625088110