[DSLX:TS][cleanup] Remove unnecessary "order" in parametric instantiation.
The parametric order is now implied by the parametrics as declared, since
parametrics are evaluated from left to right.

This is part of the general effort to move away from string-oriented parametric
environments (as keys) and towards AST-node-oriented parametric environments.
(When we have fully migrated to AST-node-oriented parametric environments we
should be able to remove ParametricSymbols as ConcreteTypeDims.)

Historical note: previously we would attempt to do more of a "constraint
satisfaction" arbitrary ordering, but the value-to-confusion ratio of that
approach is not good, so we settled on this left-to-right evaluation.

PiperOrigin-RevId: 625088126