Implement a std::clzt() operation with a tree algorithm.
Provides
  * `std::clzt_pow2()` which works on numbers with bit-lengths
    that are powers of 2
  * `std::clzt()` works for any number (the algorithm works
    on powers of two, so this appends sufficient number of
    bits first befor running; these should typically melt
    away in optimization).

To disambiguate from he builtin `clz()`, this is called
`clzt()` (t for 'tree'). Also in std lib, so FQN `std::clzt()`.

This is conceptually a recursive algorithm, but since
dslx does not support that yet, functions are manually expanded
as a proof of concept and a motivating example for #510 to
implement.

Related issues: https://github.com/google/xls/issues/510

PiperOrigin-RevId: 625827211