[DSLX:TS] "Type zipper" allows us to traverse structure of two types.
This is going to be useful for printing out prettier "type diff" error
messages that highlight where types actually have discrepancies within larger
type trees.

It has a SAX-esque callback API (if you happen to be familiar with that kind of
pattern) that invoked as you walk the two types and their common elements /
gives a signal on where they've mismatched -- we walk until there's a
structural mismatch that prevents us from walking down to more specific leaves.

Not tied into anything yet, going to build a pretty type error printer on top
of this facility.

Towards google/xls#1379

PiperOrigin-RevId: 625827631