Add a gRPC client impl of C++ Synthesizer class, and use it in delay_info_main.
A local yosys_server_main can then be used to achieve the same analysis that was previously done with YosysSynthesizer (i.e. yosys as a subprocess).

This approach will make it easier to plug synthesis into multiple tools without those tools being coupled to a particular synthesizer, its parameters, and the necessity of a local instance.

PiperOrigin-RevId: 625869523