Simplify back_propagate_range_analysis
The analysis carried around ternary values and supported inference on equality of compound data types. Neither feature was very useful and both could be quite expensive. Refactor the API to only use IntervalSets and only support bits.

This is mainly to simplify the API in preparation for a larger refactor.

PiperOrigin-RevId: 626061332