Testbuilder: specialize default printer for floating point values.
This prints the floating point values to the resolution they support,
and in addition with the %a format, that emits the exact bit pattern.

This is in particular useful to make issues in lower bits visible
(the default absl::StrCat() print just uses a default 6 digit resolution).

While at it, add a test for these previously untested functions.

PiperOrigin-RevId: 626098432