Replace child-to-parent/parent-to-child port maps.
The idea was that a visitor would use these maps to move through ports in the elaboration hierarchy, but they complicate elaboration and aren't as simplifying as I'd like. Instead, it seems cleaner for each instance to have a pointer to their parent (if it exists) and a map from `Instantiation*` (an IR construct) to `BlockInstance*` (an elaboration construct). When looking for a port in a child instance, it is easy and cheap to look up the port by name. When looking for an instantiation input/output in an parent instance, you take the instantiation of the child and search through its instantiation inputs/outputs until you find one that matches the desired port name.

Note that the search for instantiation inputs/outputs is linear as implemented. A block with sufficiently large number of ports could be slow, although in practice benchmarks show no real regression. If this becomes a performance bottleneck in the future, it is straightforward to add a map from name to `Node*` for `Instantiation`s. It seems better for the `Instantiation` IR construct to own this map than an elaboration.

PiperOrigin-RevId: 626410348