Context-sensitive Proc-State narrowing
This adds the ability to use the next-node predicates to more accurately model ranges for proc-states. By back-propagating the predicate being true we can sometimes constrain the next value significantly. By constraining all next-values for a particular proc-state we might be able to narrow it.

Currently there is no analysis of reachable values so the analysis does not work very well in cases where the predicate contains a signed compare. This is because a condition such as (signed-LT_u16 X 12) gives implied ranges for X of [[0, 12], [0x8000, 0xffff]] so all bits are unknown so no narrowing is possible. With further inductive analyses we can see through these conditions and this will be done in a later CL.

PiperOrigin-RevId: 627488817