Add more options to FIFO parameterization.
The `bypass` parameter in `FifoConfig` is not sufficient to eliminate combinational paths between push- and pop-side signals. This change adds `register_push_outputs` and `register_pop_outputs`, parameters that control if the push- or pop-side outputs are registered (breaking combinational paths).

Also, the config is getting big enough that I changed it from a struct (where some values are often unspecified) to a class with a constructor that requires all members to be specified. It seems better to be explicit everywhere, even if it's annoying (especially because there are so many `bool`s, which should be disambiguated with `/*param_name=*/`-style comments.

Ideally, I think it will be best to remove FIFO-config from channels and bind FIFO-config to channels at codegen time. See also https://github.com/google/xls/issues/1391 for more discussion.

PiperOrigin-RevId: 631217053