Ensure RangeQueryEngine does not try to create invalid Bits objects
The RangeQueryEngine when checking array-update would attempt to create bits of the indices bit-count for each possible index in the array in some circumstances. If the given indices has a bit-count which is too small to address the entire array a CHECK failure would occur when attempting to construct an out-of-range bits value.

PiperOrigin-RevId: 631844721