Don't propagate conditions through invoke nodes
The contents of a function may be side-effecting (e.g., the invoked function might contain an assert), so we can't assume any conditions for invoke nodes or propagate them to their predecessors.

This would be safe for now, due to the requirement that all tokens must eventually reach a proc's next-token node (and the fact that tokens can't be dynamically addressed by operations). However, if we break that requirement, this can break (e.g.) asserts inside functions by causing them to trigger inappropriately.

We will be breaking that requirement soon, since it's a critical step in generalizing the expressible ordering semantics in XLS IR.

PiperOrigin-RevId: 631915072