XLS IR: enable cross-activation token-ordering semantics
Before this, XLS IR did not support any way to express the constraint that "in this proc, operation A in activation 1 must execute before operation B in activation 2". All procs took a single token and returned a single token, and in order to allow pipelining of side-effecting operations, XLS codegen did not consider the next activation's token to depend on the previous activation's token.

In this change, we remove the privileged status of the special token parameter; instead, tokens behave as state parameters that happen to take up 0 bits (but cannot be optimized away). During execution (and after codegen), operations depending on a given token state parameter must wait until after the previous activation resolves the corresponding `next_value` node.

To keep their semantics the same, this change switches both DSLX & XLS[cc] from emitting procs with a token state parameter to emitting procs that materialize a new independent token at the start of each activation.

Part 1 of addressing issue #1401

PiperOrigin-RevId: 632206439